+++ sdk/cosmosdb/arm-cosmosdb/src/operations/collectionPartition.ts
@@ -130,7 +130,7 @@ const listMetricsOperationSpec: msRest.OperationSpec = {
     Parameters.collectionRid
   ],
   queryParameters: [
-    Parameters.apiVersion0,
+    Parameters.apiVersion,
     Parameters.filter0
   ],
   headerParameters: [
@@ -158,7 +158,7 @@ const listUsagesOperationSpec: msRest.OperationSpec = {
     Parameters.collectionRid
   ],
   queryParameters: [
-    Parameters.apiVersion0,
+    Parameters.apiVersion,
     Parameters.filter1
   ],
   headerParameters: [