+++ sdk/monitor/monitor-opentelemetry-exporter/src/platform/nodejs/context/context.ts
@@ -4,8 +4,8 @@
 import * as os from "os";
 import * as fs from "fs";
 import * as path from "path";
-import { Logger } from "@opentelemetry/api";
-import { ConsoleLogger, LogLevel, SDK_INFO } from "@opentelemetry/core";
+import { diag } from "@opentelemetry/api";
+import { SDK_INFO } from "@opentelemetry/core";
 
 import { ContextTagKeys } from "../../../generated";
 import { Tags } from "../../../types";
@@ -36,7 +36,6 @@ export class Context {
   private static readonly ROOT_PATH = "../../../../";
 
   constructor(
-    private _logger: Logger = new ConsoleLogger(LogLevel.WARN),
     /**
      * Path to this module's `package.json` relative to
      * `Context.ROOT_PATH`
@@ -81,7 +80,7 @@ export class Context {
         try {
           packageJson = JSON.parse(fs.readFileSync(packageJsonPathTsNode, "utf8")) as PackageJson;
         } catch (exception) {
-          this._logger.warn("Failed to load Application version", exception);
+          diag.warn("Failed to load Application version", exception);
         }
       }
 
@@ -130,7 +129,7 @@ export class Context {
         try {
           packageJson = JSON.parse(fs.readFileSync(packageJsonPathTsNode, "utf8")) as PackageJson;
         } catch (exception) {
-          this._logger.warn("Failed to load Exporter version", exception);
+          diag.warn("Failed to load Exporter version", exception);
           throw exception;
         }
       }
@@ -146,9 +145,9 @@ export class Context {
   }
 }
 
-export function getInstance(logger?: Logger, exporterPrefix?: string, appPrefix?: string): Context {
+export function getInstance(exporterPrefix?: string, appPrefix?: string): Context {
   if (!instance) {
-    instance = new Context(logger, exporterPrefix, appPrefix);
+    instance = new Context(exporterPrefix, appPrefix);
   }
   return instance;
 }