+++ sdk/monitor/monitor-opentelemetry-exporter/test/common/scenario/basic.ts
@@ -6,7 +6,7 @@ import { BasicTracerProvider } from "@opentelemetry/tracing";
 import { AzureMonitorTraceExporter } from "../../../src";
 import { Expectation, Scenario } from "./types";
 import { msToTimeSpan } from "../../../src/utils/breezeUtils";
-import { StatusCode } from "@opentelemetry/api";
+import { SpanStatusCode } from "@opentelemetry/api";
 import { FlushSpanProcessor } from "../flushSpanProcessor";
 import { delay } from "@azure/core-http";
 import { TelemetryItem as Envelope } from "../../../src/generated";
@@ -59,16 +59,16 @@ export class BasicScenario implements Scenario {
         opentelemetry.context.with(
           opentelemetry.setSpan(opentelemetry.context.active(), child1),
           () => {
-            child1.setStatus({ code: StatusCode.OK });
+            child1.setStatus({ code: SpanStatusCode.OK });
             child1.end(100);
           }
         );
 
         await delay(0);
-        child2.setStatus({ code: StatusCode.OK });
+        child2.setStatus({ code: SpanStatusCode.OK });
         child2.end(100);
 
-        root.setStatus({ code: StatusCode.OK });
+        root.setStatus({ code: SpanStatusCode.OK });
         root.end(600);
       }
     );
@@ -91,7 +91,7 @@ export class BasicScenario implements Scenario {
           version: 1,
           name: "BasicScenario.Root",
           duration: msToTimeSpan(600),
-          responseCode: "0",
+          responseCode: SpanStatusCode.OK.toString(),
           success: true,
           properties: {
             foo: "bar"
@@ -108,7 +108,7 @@ export class BasicScenario implements Scenario {
               name: "BasicScenario.Child.1",
               duration: msToTimeSpan(100),
               success: true,
-              resultCode: "0",
+              resultCode: SpanStatusCode.OK.toString(),
               properties: {
                 numbers: "123"
               }
@@ -125,7 +125,7 @@ export class BasicScenario implements Scenario {
               name: "BasicScenario.Child.2",
               duration: msToTimeSpan(100),
               success: true,
-              resultCode: "0",
+              resultCode: SpanStatusCode.OK.toString(),
               properties: {
                 numbers: "1234"
               }